functions/public_info/pyproject.toml
[project]
name = "public-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-firestore>=2.21.0",
    "google-cloud-functions>=1.20.4",
]
